tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/offscreencanvas.commit.w.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#offscreencontext2d-commit">

<script id="myWorker" type="text/worker">

function testCommitPushesContents(offscreenCanvas) {
  try {
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    ctx.commit();
    return true;
  } catch(e) {
    return false;
  }
}

function testCommitException() {
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    return true;
}

self.onmessage = function(e) {
    switch (e.data.msg) {
        case 'test1':
            self.postMessage(testCommitPushesContents(e.data.data));
            break;
        case 'test2':
            self.postMessage(testCommitException());
            break;
    }
};

</script>

<script>

function verifyPlaceholder(placeholder) {
    var canvas = document.createElement('canvas');
    canvas.width = canvas.height = 10;
    var ctx = canvas.getContext('2d');
    ctx.drawImage(placeholder, 0, 0);
    _assertPixel(canvas, 5,5, 0,255,0,255, "5,5", "0,255,0,255");
}

function makeWorker(script) {
    var blob = new Blob([script]);
    return new Worker(URL.createObjectURL(blob));
}

async_test(function(t) {
    var placeholder = document.createElement('canvas');
    placeholder.width = placeholder.height = 10;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
      assert_true(msg.data);
      verifyPlaceholder(placeholder);
    }));
    worker.postMessage({msg: 'test1', data: offscreenCanvas}, [offscreenCanvas]);
}, "Test that calling OffscreenCanvas's commit pushes its contents to its placeholder.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        assert_true(msg.data);
    }));
    worker.postMessage({msg: 'test2'});
}, "Test that calling commit on an OffscreenCanvas that is not transferred from a HTMLCanvasElement throws an exception in a worker.");

</script>
